rename ab-testing.request remove scenario from requests
