+++ src/simulation/model/request/simulation.request.ts
@@ -1,10 +1,8 @@
 import { SimulationType } from '@enums/simulation-type.enum';
-import { SimulationScenario } from '@enums/simulation-scenario.enum';
 
 import { AgentDocument } from '@db/models/agent.model';
 
 interface RunSimulationRequest {
-  scenario: SimulationScenario;
   type: SimulationType;
   name: string;
   description: string;
